requirements.txt
git+https://github.com/zodman/django-micro@fix_messages
django-widget-tweaks